requirements.txt
<<<<<<< HEAD
amqp==5.0.6
asgiref==3.3.4
attrs==21.2.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.5
Django==3.2.3
django-allauth==0.44.0
django-appconf==1.0.4
django-cookieless==1.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-guardian==2.3.0
django-rest-allauth==1.1
django-rest-authemail==2.0.6
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-sslserver==0.22
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-reset-password==1.2.8
drf-spectacular==0.17.0
drf-writable-nested==0.6.2
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
imgurpython==1.1.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.0
jsonschema==4.0.0a1
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==2.0.0
mysqlclient==2.0.3
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
sib-api-v3-sdk==6.3.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
=======
amqp>5.0.*
asgiref>3.3.4
attrs>21.2.0
billiard>3.6.4.0
celery>5.0.5
certifi>2020.12.5
cffi>1.14.5
chardet>4.0.0
click>7.1.2
click-didyoumean>0.0.3
click-plugins>1.1.1
click-repl>0.2.0
coreapi>2.3.3
coreschema>0.0.4
cryptography>3.4.7
defusedxml>0.7.1
dj-rest-auth>2.1.5
Django>3.2.3
django-allauth>0.44.0
django-appconf>1.0.4
django-cookieless>1.2
django-cors-headers>3.7.0
django-filter>2.4.0
django-guardian>2.3.0
django-rest-allauth>1.1
django-rest-authemail>2.0.6
django-rest-passwordreset>1.1.0
django-rest-swagger>2.2.0
djangorestframework
djangorestframework-simplejwt>4.6.0
drf-reset-password>1.2.8
drf-spectacular>0.17.0
drf-writable-nested>0.6.2
drf-yasg>1.20.0
gunicorn>20.1.0
idna>2.10
imgurpython>1.1.7
inflection>0.5.1
itypes>1.2.0
Jinja2>3.0.0
jsonschema>4.0.0a1
kombu>5.1.0
Markdown>3.3.4
MarkupSafe>2.0.0
mysqlclient>2.0.3
oauthlib>3.1.0
openapi-codec>1.3.2
packaging>20.9
Pillow>8.2.0
prompt-toolkit>3.0.18
psycopg2>2.8.6
pycparser>2.20
PyJWT>2.1.0
PyMySQL>1.0.2
pyparsing>2.4.7
pyrsistent>0.17.3
python-dateutil>2.8.1
python-dotenv>0.17.1
python3-openid>3.2.0
pytz>2021.1
PyYAML>5.4.1
requests>2.25.1
requests-oauthlib>1.3.0
ruamel.yaml>0.17.4
ruamel.yaml.clib>0.2.2
sib-api-v3-sdk>6.3.1
simplejson>3.17.2
six>1.16.0
sqlparse>0.4.1
swagger-spec-validator>2.7.3
uritemplate>3.0.1
urllib3>1.26.4
vine>5.0.0
wcwidth>0.2.5
>>>>>>> 7a43b480b73ca33603003f360004c0b9377bec6e
